add nginx folder, remove colors folder, update request styles

+++ src/ethereum/contracts/CloneFactory.sol
@@ -15,7 +15,7 @@ contract CloneFactory {
         baseImplementation = address(new Campaign());
     }
 
-    function createCampaign(uint40 _minimum, address _creator) external {
+    function createCampaign(uint _minimum, address _creator) external {
         address _newCampaign = Clones.clone(baseImplementation);
         Campaign(_newCampaign).initialize(_minimum, _creator);
         campaignContracts.push(_newCampaign);
